Filters have been fixed and now correctly search through a stored defaultProducts state.
Filters are now contained in small dropdown menus, they still need to be arranged correctly.

Added an "Action" field to the table, this allows the admin to edit/delete product.
^
possible idea, when the mouse hovers over a row, darken that row and when its clicked open a new page with all the details and actions for that product